Refactor to extract logic
